requirements.txt
asgiref==3.7.2
black==23.7.0
cfgv==3.3.1
click==8.1.6
distlib==0.3.7
Django==4.2.3
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
exceptiongroup==1.1.2
filelock==3.12.2
identify==2.5.26
iniconfig==2.0.0
mirakuru==2.5.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.9.1
pluggy==1.2.0
port-for==0.7.1
pre-commit==3.3.3
psutil==5.9.5
psycopg==3.1.9
psycopg2==2.9.6
PyJWT==2.8.0
pytest==7.4.0
pytest-django==4.5.2
pytest-postgresql==5.0.0
pytz==2023.3
PyYAML==6.0.1
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
virtualenv==20.24.2
